Move Order controllers to same namespace
