+++ lib/Alchemy/Phrasea/ControllerProvider/Prod/Order.php
@@ -13,8 +13,8 @@ namespace Alchemy\Phrasea\ControllerProvider\Prod;
 
 use Alchemy\Phrasea\Application as PhraseaApplication;
 use Alchemy\Phrasea\Controller\LazyLocator;
-use Alchemy\Phrasea\Controller\Prod\OrderController;
 use Alchemy\Phrasea\ControllerProvider\ControllerProviderTrait;
+use Alchemy\Phrasea\Order\Controller\ProdOrderController;
 use Alchemy\Phrasea\Order\OrderBasketProvider;
 use Alchemy\Phrasea\Order\OrderValidator;
 use Silex\Application;
@@ -39,7 +39,7 @@ class Order implements ControllerProviderInterface, ServiceProviderInterface
         });
 
         $app['controller.prod.order'] = $app->share(function (PhraseaApplication $app) {
-            return (new OrderController($app))
+            return (new ProdOrderController($app))
                 ->setDispatcher($app['dispatcher'])
                 ->setEntityManagerLocator(new LazyLocator($app, 'orm.em'))
                 ->setUserQueryFactory(new LazyLocator($app, 'phraseanet.user-query'))
